Provide a simple API for fetching stats
The API endpoint at /stats will return information about the most viewed
pages, the appearance in search results and downloads of data files.

Making a simple GET request to /stats will return the top 10 most
viewed/searched/downloaded items.  The following parameters will provide
a mechanism for filtering the results.

    orgs - A comma separated list of organisation IDs that we want to filter on
    year - Only return results for the specified year
    month - When present with year, only return results for the
    specified month/year
    count - The number of records to return
    stat - one of view/download/search to filter results to only those
    actions

It's entirely possible we may need to change this to a POST if we need
to support sending lots of organisation IDs